Low Level format optional choose where the characters print
If for example you are doing a display or the like, maybe you don't want the
32 dots to print on Serial, but maybe to your display.  So add optional parameter, to allow you to pass anything derived from class Print.
Defaults to Serial.

Also a few #ifdef LATER
in the NAND code for code that was setting up a command object but than not do anything with them.  This removed compiler warnings.

@mjs513 @Defragster - you might want to check if I missed anything.
I did run MTP with these changes and did to a format of a LittleFS RAM disk and the dots still showed up on debug window.