RAM reset on any restart, cache flush not needed
the: arm_dcache_flush_delete()  covers ALL the cache over the entire DMA_MEM alloc for a RAM drive there. And serves no purpose if the MCU handles the cache, and the drive is always reset across ANY restart. This was only added to support warm start - and in any case should not have called for _delete preventing re-use.