Add Cypress  and Fuji 2MBit Fram chips...
@PaulStoffregen  @mjs513 @Defragster

I thought I would try a few Fram chips

A lot of them were out of stock, but found 2 different 2mbit chips, so got a few.

I added them to the list and they appear to work.  I brought them up in MTP sketch and was able to upload a file to them and down download it...

Not sure if you want it for this release, but...