Accommodate larger PSRAM parts
See discussion at https://forum.pjrc.com/index.php?threads/additional-psram-id-that-works-plus-goodies.76942/

Rather than using a fixed address for the Teensy 4.1
onboard QSPI Flash, this commit uses an address
computed in a revised startup.c, based on the fitted
PSRAM size.

If this version of LittleFS is used with an older
version of Teensyduino then the required symbol
will not be defined, so it contains a weak definition
of that symbol with the original hard-coded address,
so LittleFS should continue to work.